Add "ds" attribute to results vectors
Add dataset ("ds") attribute to the output vectors. The "ds" value is determined from the patch filename, this only works for the situation we had, when we were creating results for dealing with false positives. This should be changed in future!

It also dissolves the result based on the dataset name, meaning if there are 2 overlapping vectors from the same dataset, the results will be dissolved.